Ensure that the plugin does not get applied twice
This plugin may be applied manually in order to make use of the
bspConfig generation.

It may also be applied via init scripts in the case of commands run via
the Gradle Tooling API.

Applying the plugin twice causes errors when trying to register tasks
and configurations that already exist.